pyproject.toml
[project]
name = "bot_email"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "kaubarros-24", email = "martinsbarroskaua@gmail.com"},
]
dependencies = ["django>=5.1.4", "djangorestframework>=3.15.2", "passage-identity>=2.5.1", "simplejwt>=2.0.1", "djangorestframework-simplejwt>=5.3.1", "drf-spectacular>=0.28.0", "python-dotenv>=1.0.1", "reportlab>=4.2.5", "google-generativeai>=0.8.3"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
r = "python3 manage.py runserver"
makemigrations = "python3 manage.py makemigrations"
migrate = "python3 manage.py migrate"
newApp = "django-admin startapp"
shellp = "python3 manage.py shell"